+++ src/paco/cftemplates/__init__.py
@@ -50,3 +50,4 @@ from paco.cftemplates.ecr import ECRRepository
 from paco.cftemplates.iamuser import IAMUser
 from paco.cftemplates.pinpointapplication import PinpointApplication
 from paco.cftemplates.cognito import CognitoUserPool, CognitoIdentityPool
+from paco.cftemplates.dynamodb import DynamoDB